Allow configuring the default lab size
Add a new `controller.config.lab.defaultSize` configuration option
that sets the default lab size explicitly. If it is not set, or is
set to a size the user's quota does not allow them to spawn, the
default remains the first-listed size that the user's quota allows
them to spawn.