Rewrite build system to use uv.lock files
Reorganize the build system into multiple workspaces so that all
parts of the build can use `uv.lock` files and nox-uv for simpler
`noxfile.py` files.

Create a (for now, dummy) top-level nublado package to serve as the
workspace root, and include inithome and purger in that namespace
since there is no pressing need for them to have separate virtual
environments or frozen dependencies. Maintain the controller and
the Nublado client as separate projects with their own `uv.lock`
files since the controller has heavier dependencies and the client
may need to support a wider range of Python versions.

Turn the hub directory into a (currently dummy) Python package named
nublado-jupyterhub so that it can manage its own dependencies. Make
that a workspace root and move authenticator and spawner into a
plugins directory inside that workspace.

Merge the test suites of inithome and purger, and of authenticator
and spawner, to follow the rule of one test suite per workspace and
dev dependency management at the root of the workspace.

Use recursive nox to build all of these projects together. This
allows use of nox-uv at each level and far simpler `noxfile.py` files.
It would have been possible to use non-recursive nox, but the
`noxfile.py` file would have been much more complex.

Use uv sync to install dependencies, except for the JupyterHub image.
In that case, export the `uv.lock` file to a `requirements.txt` so
that the packages can still be installed in the system paths and
override the packages that come with the Zero to JupyterHub container.
For the controller, inithome, and purger images, use our current
virtual environment management approach for service builds.