Add dependency version bounds
Add lower version bounds on all of the main dependencies to ensure
that uv or pip doesn't install an ancient version as a way of
satisfying a new dependency. Add `required-environments` as a
setting due to an error message received when updating to uv 0.9.0
to prevent upgrading to a version that doesn't have a wheel available
for the platforms we care about.