Make JupyterLab build conditional
The nublado-jupyterlab-base image takes a while to build and is
rather large, and is also isolated from the rest of Nublado. Only
build it if one of the inputs that affects it has changed, or on
workflow dispatch.

Nublado uses a standard naming convention for its images, so
construct the image names using GitHub Actions templating rather than
setting them in the matrix. This keeps the names of the jobs in
GitHub Actions much shorter and more readable and should produce the
same results.