Add Python client API documentation
Generation Python client API documentation and include it in the
documentation site. Convert the top-level dependencies used for
documentation to editable dependencies, since otherwise old versions
of the dependencies are cached and not updated even if the nox
virtual environments are recreated. Add some more markup to the
Nublado client exceptions to force buggy Sphinx to do the right thing.