Overhaul Nublado client documentation
Rework the Nublado client documentation to hopefully be clearer.
Replace lots of literal strings with proper cross-references to the
relevant classes, methods, and modules.

Simplify the examples of how to use the Nublado client, and shorten
the output from one of the tests. Remove the complicated code to
compare the output of two ways of running notebooks, since it doesn't
make a good example.