Remove loop around spawn progress in client
A long time ago, the server-sent events endpoint we watched for
spawn progress events would sometimes close after the first message
and require a retry. Remove the workaround for this bug in the hope
that this was an old kubespawner bug that should not happen with
the new Nublado spawner plugin.