Move and separate the JupyterLab session code
Move the code to manage the JupyterLab sessions and WebSocket into
a separate internal `_session` module, and separate the session
itself, with the `run_python` method, from the session manager that
provides the context manager. This allows removal of a check for
whether the context manager has been opened, since errors like that
are now caught by mypy.