Significantly refactor the mock API
Rework how code execution in the mock is handled to simplify the
internals. Replace all public object attributes with methods to
perform the same tasks, with proper documentation. Rename
`register_extension_result` to `register_notebook_result` because
the caller doesn't care how the notebook is run. Allow the argument
to `register_python_result` to be an exception, which is raised and
handled as an error, and remove the special case for the code
`long_error_for_test()`.