Move notebook execution test to client tests
Move the test for whole notebook execution out of the mock tests
into the client tests. Add a helper function to read a test data
file and use it. Simplify the result model by adding another default
value, and clean up the error model to use more sensible field names
and use a validation alias to support the existing names.