Remove complex notebook normalization in mock
Previously, when registering an execution result for a notebook, the
mock did complex normalization on the contents of the notebook to
determine a key with which to look up registered results. This is all
unnecessary for tests since the Nublado client passes the text of the
notebook through without modification. Just use the raw text of the
notebook as the key.