Update Python version, fix purger configuration
The configuration for the purger declared a camel-case alias
generator, but with previous versions of Pydantic it wasn't fully
enforced and snake-case keys were still accepted. (It's not entirely
clear why this was.)

Switch the purger configuration fully over to camel-case and add
aliases for the relevant settings so that the previous environment
variables still work for configuration overrides. Update the test
configuration to also use camel-case.