Fix periodic testing with old Python versions
Periodic tests with old Python versions were failing because uv
tried to use the old Python to do all lock file updates, and that
fails for components that don't support that version. Always use
the current Python version for lock file updates, and then install
the old Python version and use it to run tests.